\name{as.matrix.cor.matrix}
\alias{as.matrix.cor.matrix}
\title{as.matrix method}
\description{
  as matrix
}
\usage{
\method{as.matrix}{cor.matrix}(x,...)
}
\arguments{
  \item{x}{Object of class \code{"cor.matrix"}.}
  \item{\dots}{further arguments. unsued}  
}
\value{
  a matrix
} 
