 \name{contin.tests.to.table}
\alias{contin.tests.to.table}
\title{
  contin.tests.to.table
}
\description{
  Makes a nice table out of a \code{contin.tests} object
}
\usage{
contin.tests.to.table(tests,test.digits=3,...)
}
\arguments{
  \item{tests}{a \code{contin.tests} object}
  \item{test.digits}{The number of digits to round to}
  \item{\dots}{other paramaters}
}
\value{
  A nice table
}
