\name{multi.test}
\alias{multi.test}
\title{multi.test}
\description{
  Creates a table from a list of htests
}
\usage{
multi.test(tests)
}
\arguments{
  \item{tests}{A named list of htest objects representing the same test applied to a number of different conditions or variables.}
}
\value{
  A \code{multi.test} object, representing a table of the \code{htest} objects.
}  
