\name{plot.cor.matrix}
\alias{plot.cor.matrix}
\title{Plot method}
\description{
  Produces a circle plot for an object of class \code{"plot.cor.matrix"}
}
\usage{
\method{plot}{cor.matrix}(x,y=NULL,size=10,...)
}
\arguments{
  \item{x}{Object of class \code{"cor.matrix"}.}
  \item{y}{unused}
  \item{size}{maximum radius size}
  \item{\dots}{further arguments. unsued}  
}
\value{
  a ggplot object
} 
