\name{print.contin.tests}
\alias{print.contin.tests}
\title{Print method }
\description{
  Print object of class \code{"contin.tests"} in nice layout.
}
\usage{
\method{print}{contin.tests}(x,test.digits, \dots)
}
\arguments{
  \item{x}{Object of class \code{"contin.tests"}.}
  \item{test.digits}{Number of digits to be printed} 
  \item{\dots}{further arguments to be passed to or from methods.}  
}

\value{
  none
}
