\name{print.freq.table}
\alias{print.freq.table}
\title{Print method}
\description{
  Print object of class \code{"freq.table"} in nice layout.
}
\usage{
\method{print}{freq.table}(x,\dots)
}
\arguments{
  \item{x}{Object of class \code{"freq.table"}.}
  \item{\dots}{further arguments}  
}

\value{
  none
}
 