\name{print.multi.test}
\alias{print.multi.test}
\title{Print method}
\description{
  Print object of class \code{"multi.test"} in nice layout.
}
\usage{
\method{print}{multi.test}(x,...)
}
\arguments{
  \item{x}{Object of class \code{"multi.test"}.}
  \item{\dots}{further arguments}  
}
\value{
  none
}   